Cleanup unused return from parseData. Add looping ADR command to try to clear iPM device and get it to respond.
When installed on the GV (as opposed to in the lab), on power up the iPM frequently comes up in a hung state (returns nothing in response to sent commands. The working theory is that there is some junk on the port that is corrupting commands being sent. Attempt to clear this by repeatedly sending the ADR and VER? commands up to 10 times.